Introducing MemeSharer user authentication system! Now, upon user sign in, they are issued a signed and encrypted JWT which carries all the permissions they have access to and their user information. This token is stored in the user's local storage and is sent as a request header to any backend API endpoint which requires validation of the user's authenticity or confirmation that the user has the permissions to perform certain queries/mutations. Along with this tokenized access system comes the perms or permissions microservice. This microservice establishes role-based user authenication. Users can be assigned multiple roles, roles can be assigned multiple permissions, and permissions state the backend api function name to which a user might need access to. There is much more to be discovered in this sizable update, as it will be the core backbone upon which all authentication of users is predicated. Feel free to message me if you have any questions about the particulars of this system. A final note: I would say that this update is the point at which this application becomes quite complex. For any developer wishing to simply understand the basics of a social network built on the MERN stack, please pull from the prior commit to this one. Past this point, the application requires much more in depth security protocols, such as multiple RSA key pair requirements that coincide with a deep understanding of encryption. The app also exponentiates in complexity with the introduction of permissions microservice, to which the developer must fully understand role-based authentication. For these reasons, I encourage anyone to pull from this or future commits if they are willing to put in more time into the setup for a more secure and scalable application. Best, Anthony Pitts

+++ app/services/perms/app/graphql/resolvers/perms-mutations.js
@@ -7,6 +7,9 @@ require("module-alias/register");
 const { handleErrors } = require("@config/handleGQLErrors");
 const { logSuccess, logFail } = require("@config/logger");
 const { logger } = require("@config/logger.js");
+const { AuthenticateToken } = require('../../../lib/AuthenticateToken')
+const { addReqHeaders } = require('../../../lib/addReqHeaders.js')
+
 
 const { createApolloFetch } = require("apollo-fetch");
 // Models
@@ -20,43 +23,82 @@ const validateInput = require("../../validation/validateInput.js");
 // CREATE MUTATIONS
 //==============================================================================
 
-const createRoleMutation = async (_, { input }) => {
+const createRoleMutation = async (_, { input }, { req }) => {
   try {
+
     validateInput(input, "createRoleMutation");
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //'invalid user auth token or not signed in'
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to create a role
+    if(!authTokenData.hasPermission(`createRole`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to create a role`)
+
     let role = await Role.findOne({ roleName: input.roleName });
     handleErrors.duplicateRoleNameOnCreate(role);
+
     const newRole = new Role({
       roleName: input.roleName
     });
     const saveNewRole = await newRole.save();
+
     logSuccess("createRoleMutation", newRole);
     return saveNewRole;
+
   } catch (err) {
     logFail("createRoleMutation", err);
-    throw err;
   }
 };
 
-const createPermissionMutation = async (_, { permName }) => {
+const createPermissionMutation = async (_, { permName }, { req }) => {
   try {
+
     validateInput({ permName: permName }, "createPermissionMutation");
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to create a permission
+    if(!authTokenData.hasPermission(`createPermission`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to create a permission`)
+
     let permission = await Permission.findOne({ permName: permName });
     handleErrors.duplicatePermNameOnCreate(permission);
     const newPermission = new Permission({
       permName: permName
     });
     const saveNewPermission = newPermission.save();
+
     logSuccess("createPermissionMutation", saveNewPermission);
     return saveNewPermission;
+
   } catch (err) {
     logFail("createPermissionMutation", err);
-    throw err;
   }
 };
 
-const addRolePermissionMutation = async (_, { input }) => {
+const addRolePermissionMutation = async (_, { input }, { req }) => {
   try {
+
     validateInput(input, "addRolePermissionMutation");
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to add role permission
+    if(!authTokenData.hasPermission(`addRolePermission`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to add role permission`)
+
     var permission = await Permission.findById(input.permID);
     var role = await Role.findById(input.roleID);
     var rolePerm = await RolePermission.findOne({
@@ -71,17 +113,30 @@ const addRolePermissionMutation = async (_, { input }) => {
       permID: input.permID
     });
     let saveNewRolePermission = await newRolePermission.save();
+
     logSuccess("addRolePermissionMutation", saveNewRolePermission);
     return saveNewRolePermission;
+
   } catch (err) {
     logFail("addRolePermissionMutation", err);
-    throw err;
   }
 };
 
-const addUserRoleMutation = async (_, { input }) => {
+const addUserRoleMutation = async (_, { input }, { req }) => {
   try {
+
     validateInput(input, "addUserRoleMutation");
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to add user role
+    if(!authTokenData.hasPermission(`addUserRole`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to add user role`)
+
     var role = await Role.findById(input.roleID);
     var userRole = await UserRole.findOne({
       roleID: input.roleID,
@@ -94,19 +149,31 @@ const addUserRoleMutation = async (_, { input }) => {
       userID: input.userID
     });
     let saveNewUserRole = await newUserRole.save();
+
     logSuccess("addUserRoleMutation", saveNewUserRole);
     return saveNewUserRole;
+
   } catch (err) {
     logFail("addUserRoleMutation", err);
-    throw err;
   }
 };
 
 //==============================================================================
 // DELETE MUTATIONS
 //==============================================================================
-const removePermissionFromRoleMutation = async (_, { input }) => {
+const removePermissionFromRoleMutation = async (_, { input }, { req }) => {
   try {
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to remove permission from role
+    if(!authTokenData.hasPermission(`removePermissionFromRole`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to remove permission from role`)
+
     var rolePermissionToDelete = await RolePermission.deleteOne({
       roleID: input.roleID,
       permissionID: input.permissionID
@@ -119,8 +186,19 @@ const removePermissionFromRoleMutation = async (_, { input }) => {
   }
 };
 
-const deleteRoleMutation = async (_, { id }) => {
+const deleteRoleMutation = async (_, { id }, { req }) => {
   try {
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to delete a role
+    if(!authTokenData.hasPermission(`deleteRole`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to delete a role`)
+
     //check if role exists
     var role = await Role.findById(id);
     if (!role) {
@@ -129,9 +207,13 @@ const deleteRoleMutation = async (_, { id }) => {
     }
     await RolePermission.deleteMany({ roleID: id });
     await Role.findByIdAndDelete(id);
+    
     var fetch = createApolloFetch({
       uri: `${process.env.ssl}://${process.env.website_name}:${process.env.gatewayms_port}/gateway`
     });
+    //sets the authorization request header
+    addReqHeaders(fetch, req.headers.authorization);
+
     await fetch({
       query: `
       mutation deleteUserRolesByRoleID($id: String!){
@@ -150,9 +232,19 @@ const deleteRoleMutation = async (_, { id }) => {
   }
 };
 
-const removeUserRoleMutation = async (_, { input }) => {
+const removeUserRoleMutation = async (_, { input }, { req }) => {
   try {
     validateInput(input, "removeUserRoleMutation");
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to remove user role
+    if(!authTokenData.hasPermission(`removeUserRole`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to remove a user role`)
 
     let deleted = await UserRole.deleteMany({
       roleID: input.roleID,
@@ -167,8 +259,19 @@ const removeUserRoleMutation = async (_, { input }) => {
   }
 };
 
-const deletePermissionMutation = async (_, { permissionID }) => {
+const deletePermissionMutation = async (_, { permissionID }, { req }) => {
   try {
+
+    //Get the signed-in user's decrypted auth token payload
+    const authTokenData = new AuthenticateToken(req);
+
+    //invalid user auth token or not signed in
+    if(authTokenData.errors) return handleErrors.invalidAuthToken(authTokenData.errors)
+
+    //if signed in user doesn't have access to delete a permission
+    if(!authTokenData.hasPermission(`deletePermission`)) 
+      return handleErrors.permissionDenied(`Signed in user doesn't have access to delete a permission`)
+
     var perm = await Permission.findById(permissionID);
     if (!perm) {
       logger.error("Permission doesn't exist");
